pyproject.toml
[tool.poetry]
name = "medhelper"
version = "0.1.0"
description = "Medical card for patients and doctors with history of diseases and analyzes"
authors = ["Ostap Bodnar <ostapbod674@gmail.com>", 
           "Nadya Horishna <nadyahorishna2002@gmail.com>",
           "Vitalii Shkliarov shkliarov.vitalik@gmail.com",
           "Ivan Dumych ivan69175@gmail.com",
           "Duda Oleksandr olexandrduda@gmail.com",
           "Ostap Matviiv ostap20021906@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
psycopg2-binary = "^2.8.6"
python-dotenv = "^0.15.0"
djangorestframework = "^3.12.2"
Pillow = "^8.0.1"
django-cors-headers = "^3.5.0"
djangorestframework-simplejwt = "^4.6.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
